Final update to monitor.cpp file
There are a few print statements that need to be adjusted, you should they have just an "x" in them. Please adjust those when you have a chance tomorrow, I have a few examples in there with using the "threadNum" integer for thread identification. The algorithm for tracking the queue rotation is donated with qTrack for the Consumer, and prodTrack for the producer. 
Please double check to make sure I am not missing any global variables